requirements.txt
falcon
pydantic
pydantic-settings
psycopg[binary]
elasticsearch
boto3
aiofiles
python-dotenv
asyncpg
uvicorn
gunicorn
aiohttp